app/services/ai_inference_service/pyproject.toml
[project]
name = "ai-inference-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "ctranslate2>=4.6.0",
    "sentencepiece>=0.2.1",
    "pydantic-extra-types>=2.10.5",
    "pydantic-settings>=2.10.1",
]
